+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
 * Improved typings for module compositer
+* A browser won't have 2 tabs opened by default
 
 2.4.0 / 2019-07-15
 ==================